web_app/requirements.txt
acachecontrol==0.3.5
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
dash==2.9.1
dash-ace==0.2.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dashvis==0.1.3
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-oidc==1.4.0
Flask-OpenID==1.3.0
fonttools==4.39.2
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
huggingface-hub==0.13.1
humanfriendly==10.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonlines==3.1.0
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
networkx==3.0
numpy==1.24.2
oauth2client==4.1.3
okta==0.0.4
okta-jwt-verifier==0.2.3
openai==0.27.2
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
retry==0.9.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
selenium==4.8.3
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tenacity==8.2.2
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.26.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
